Use the new collection TCK
Moved tests from APITest(api_test.js) to CollectionTest(collection_test.js)
