+++ src/test/resources/testmodel-js/test_interface.js
@@ -456,896 +456,12 @@ var TestInterface = function(j_val) {
   /**
 
    @public
-   @param dataObjects {Array.<Object>} 
-   */
-  this.methodWithListOfDataObjectsParam = function(dataObjects) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'object' && __args[0] instanceof Array) {
-      j_testInterface["methodWithListOfDataObjectsParam(java.util.List)"](utils.convParamListDataObject(dataObjects, function(json) { return new TestDataObject(json); }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param dataObjects {Array.<Object>} 
-   */
-  this.methodWithSetOfDataObjectsParam = function(dataObjects) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'object' && __args[0] instanceof Array) {
-      j_testInterface["methodWithSetOfDataObjectsParam(java.util.Set)"](utils.convParamSetDataObject(dataObjects, function(json) { return new TestDataObject(json); }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
    @param dataObject {Object} 
    */
   this.methodWithNullDataObjectParam = function(dataObject) {
     var __args = arguments;
     if (__args.length === 1 && typeof __args[0] === 'object') {
       j_testInterface["methodWithNullDataObjectParam(io.vertx.codegen.testmodel.TestDataObject)"](dataObject != null ? new TestDataObject(new JsonObject(JSON.stringify(dataObject))) : null);
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listString {Array.<string>} 
-   @param listByte {Array.<number>} 
-   @param listShort {Array.<number>} 
-   @param listInt {Array.<number>} 
-   @param listLong {Array.<number>} 
-   @param listJsonObject {Array.<Object>} 
-   @param listJsonArray {Array.<todo>} 
-   @param listVertxGen {Array.<RefedInterface1>} 
-   @param listDataObject {Array.<Object>} 
-   @param listEnum {Array.<Object>} 
-   */
-  this.methodWithListParams = function(listString, listByte, listShort, listInt, listLong, listJsonObject, listJsonArray, listVertxGen, listDataObject, listEnum) {
-    var __args = arguments;
-    if (__args.length === 10 && typeof __args[0] === 'object' && __args[0] instanceof Array && typeof __args[1] === 'object' && __args[1] instanceof Array && typeof __args[2] === 'object' && __args[2] instanceof Array && typeof __args[3] === 'object' && __args[3] instanceof Array && typeof __args[4] === 'object' && __args[4] instanceof Array && typeof __args[5] === 'object' && __args[5] instanceof Array && typeof __args[6] === 'object' && __args[6] instanceof Array && typeof __args[7] === 'object' && __args[7] instanceof Array && typeof __args[8] === 'object' && __args[8] instanceof Array && typeof __args[9] === 'object' && __args[9] instanceof Array) {
-      j_testInterface["methodWithListParams(java.util.List,java.util.List,java.util.List,java.util.List,java.util.List,java.util.List,java.util.List,java.util.List,java.util.List,java.util.List)"](utils.convParamListBasicOther(listString), utils.convParamListByte(listByte), utils.convParamListShort(listShort), utils.convParamListBasicOther(listInt), utils.convParamListLong(listLong), utils.convParamListJsonObject(listJsonObject), utils.convParamListJsonArray(listJsonArray), utils.convParamListVertxGen(listVertxGen), utils.convParamListDataObject(listDataObject, function(json) { return new TestDataObject(json); }), utils.convParamListEnum(listEnum, function(val) { return Packages.io.vertx.codegen.testmodel.TestEnum.valueOf(val); }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param setString {Array.<string>} 
-   @param setByte {Array.<number>} 
-   @param setShort {Array.<number>} 
-   @param setInt {Array.<number>} 
-   @param setLong {Array.<number>} 
-   @param setJsonObject {Array.<Object>} 
-   @param setJsonArray {Array.<todo>} 
-   @param setVertxGen {Array.<RefedInterface1>} 
-   @param setDataObject {Array.<Object>} 
-   @param setEnum {Array.<Object>} 
-   */
-  this.methodWithSetParams = function(setString, setByte, setShort, setInt, setLong, setJsonObject, setJsonArray, setVertxGen, setDataObject, setEnum) {
-    var __args = arguments;
-    if (__args.length === 10 && typeof __args[0] === 'object' && __args[0] instanceof Array && typeof __args[1] === 'object' && __args[1] instanceof Array && typeof __args[2] === 'object' && __args[2] instanceof Array && typeof __args[3] === 'object' && __args[3] instanceof Array && typeof __args[4] === 'object' && __args[4] instanceof Array && typeof __args[5] === 'object' && __args[5] instanceof Array && typeof __args[6] === 'object' && __args[6] instanceof Array && typeof __args[7] === 'object' && __args[7] instanceof Array && typeof __args[8] === 'object' && __args[8] instanceof Array && typeof __args[9] === 'object' && __args[9] instanceof Array) {
-      j_testInterface["methodWithSetParams(java.util.Set,java.util.Set,java.util.Set,java.util.Set,java.util.Set,java.util.Set,java.util.Set,java.util.Set,java.util.Set,java.util.Set)"](utils.convParamSetBasicOther(setString), utils.convParamSetByte(setByte), utils.convParamSetShort(setShort), utils.convParamSetBasicOther(setInt), utils.convParamSetLong(setLong), utils.convParamSetJsonObject(setJsonObject), utils.convParamSetJsonArray(setJsonArray), utils.convParamSetVertxGen(setVertxGen), utils.convParamSetDataObject(setDataObject, function(json) { return new TestDataObject(json); }), utils.convParamSetEnum(setEnum, function(val) { return Packages.io.vertx.codegen.testmodel.TestEnum.valueOf(val); }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param mapString {Array.<string>} 
-   @param mapByte {Array.<string>} 
-   @param mapShort {Array.<string>} 
-   @param mapInt {Array.<string>} 
-   @param mapLong {Array.<string>} 
-   @param mapJsonObject {Array.<string>} 
-   @param mapJsonArray {Array.<string>} 
-   @param mapVertxGen {Array.<string>} 
-   */
-  this.methodWithMapParams = function(mapString, mapByte, mapShort, mapInt, mapLong, mapJsonObject, mapJsonArray, mapVertxGen) {
-    var __args = arguments;
-    if (__args.length === 8 && (typeof __args[0] === 'object' && __args[0] != null) && (typeof __args[1] === 'object' && __args[1] != null) && (typeof __args[2] === 'object' && __args[2] != null) && (typeof __args[3] === 'object' && __args[3] != null) && (typeof __args[4] === 'object' && __args[4] != null) && (typeof __args[5] === 'object' && __args[5] != null) && (typeof __args[6] === 'object' && __args[6] != null) && (typeof __args[7] === 'object' && __args[7] != null)) {
-      j_testInterface["methodWithMapParams(java.util.Map,java.util.Map,java.util.Map,java.util.Map,java.util.Map,java.util.Map,java.util.Map,java.util.Map)"](mapString, utils.convParamMapByte(mapByte), utils.convParamMapShort(mapShort), mapInt, utils.convParamMapLong(mapLong), utils.convParamMapJsonObject(mapJsonObject), utils.convParamMapJsonArray(mapJsonArray), utils.convParamMapVertxGen(mapVertxGen));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listStringHandler {function} 
-   @param listIntHandler {function} 
-   @param setStringHandler {function} 
-   @param setIntHandler {function} 
-   */
-  this.methodWithHandlerListAndSet = function(listStringHandler, listIntHandler, setStringHandler, setIntHandler) {
-    var __args = arguments;
-    if (__args.length === 4 && typeof __args[0] === 'function' && typeof __args[1] === 'function' && typeof __args[2] === 'function' && typeof __args[3] === 'function') {
-      j_testInterface["methodWithHandlerListAndSet(io.vertx.core.Handler,io.vertx.core.Handler,io.vertx.core.Handler,io.vertx.core.Handler)"](function(jVal) {
-      listStringHandler(jVal);
-    }, function(jVal) {
-      listIntHandler(jVal);
-    }, function(jVal) {
-      setStringHandler(utils.convReturnSet(jVal));
-    }, function(jVal) {
-      setIntHandler(utils.convReturnSet(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   */
-  this.methodWithHandlerAsyncResultListString = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultListString(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        handler(ar.result(), null);
-      } else {
-        handler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   */
-  this.methodWithHandlerAsyncResultListInteger = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultListInteger(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        handler(ar.result(), null);
-      } else {
-        handler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   */
-  this.methodWithHandlerAsyncResultSetString = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultSetString(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        handler(utils.convReturnSet(ar.result()), null);
-      } else {
-        handler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   */
-  this.methodWithHandlerAsyncResultSetInteger = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultSetInteger(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        handler(utils.convReturnSet(ar.result()), null);
-      } else {
-        handler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerListVertxGen = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerListVertxGen(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetVertxGen(jVal, RefedInterface1));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerSetVertxGen = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerSetVertxGen(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetVertxGen(jVal, RefedInterface1));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerListAbstractVertxGen = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerListAbstractVertxGen(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetVertxGen(jVal, RefedInterface2));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerSetAbstractVertxGen = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerSetAbstractVertxGen(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetVertxGen(jVal, RefedInterface2));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerListJsonObject = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerListJsonObject(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetJson(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerListNullJsonObject = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerListNullJsonObject(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetJson(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerListComplexJsonObject = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerListComplexJsonObject(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetJson(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerSetJsonObject = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerSetJsonObject(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetJson(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerSetNullJsonObject = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerSetNullJsonObject(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetJson(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerSetComplexJsonObject = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerSetComplexJsonObject(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetJson(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerListJsonArray = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerListJsonArray(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetJson(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerListNullJsonArray = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerListNullJsonArray(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetJson(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerListComplexJsonArray = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerListComplexJsonArray(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetJson(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerSetJsonArray = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerSetJsonArray(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetJson(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerSetNullJsonArray = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerSetNullJsonArray(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetJson(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param setHandler {function} 
-   */
-  this.methodWithHandlerSetComplexJsonArray = function(setHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerSetComplexJsonArray(io.vertx.core.Handler)"](function(jVal) {
-      setHandler(utils.convReturnListSetJson(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerListDataObject = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerListDataObject(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetDataObject(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerListNullDataObject = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerListNullDataObject(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetDataObject(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param setHandler {function} 
-   */
-  this.methodWithHandlerSetDataObject = function(setHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerSetDataObject(io.vertx.core.Handler)"](function(jVal) {
-      setHandler(utils.convReturnListSetDataObject(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param setHandler {function} 
-   */
-  this.methodWithHandlerSetNullDataObject = function(setHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerSetNullDataObject(io.vertx.core.Handler)"](function(jVal) {
-      setHandler(utils.convReturnListSetDataObject(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerListEnum = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerListEnum(io.vertx.core.Handler)"](function(jVal) {
-      listHandler(utils.convReturnListSetEnum(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param setHandler {function} 
-   */
-  this.methodWithHandlerSetEnum = function(setHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerSetEnum(io.vertx.core.Handler)"](function(jVal) {
-      setHandler(utils.convReturnListSetEnum(jVal));
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultListVertxGen = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultListVertxGen(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetVertxGen(ar.result(), RefedInterface1), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultSetVertxGen = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultSetVertxGen(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetVertxGen(ar.result(), RefedInterface1), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultListAbstractVertxGen = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultListAbstractVertxGen(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetVertxGen(ar.result(), RefedInterface2), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultSetAbstractVertxGen = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultSetAbstractVertxGen(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetVertxGen(ar.result(), RefedInterface2), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultListJsonObject = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultListJsonObject(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetJson(ar.result()), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultListNullJsonObject = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultListNullJsonObject(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetJson(ar.result()), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultListComplexJsonObject = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultListComplexJsonObject(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetJson(ar.result()), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultSetJsonObject = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultSetJsonObject(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetJson(ar.result()), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultSetNullJsonObject = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultSetNullJsonObject(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetJson(ar.result()), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultSetComplexJsonObject = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultSetComplexJsonObject(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetJson(ar.result()), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultListJsonArray = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultListJsonArray(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetJson(ar.result()), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultListNullJsonArray = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultListNullJsonArray(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetJson(ar.result()), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultListComplexJsonArray = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultListComplexJsonArray(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetJson(ar.result()), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultSetJsonArray = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultSetJsonArray(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetJson(ar.result()), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultSetNullJsonArray = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultSetNullJsonArray(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetJson(ar.result()), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultSetComplexJsonArray = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultSetComplexJsonArray(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetJson(ar.result()), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultListDataObject = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultListDataObject(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetDataObject(ar.result()), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultListNullDataObject = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultListNullDataObject(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetDataObject(ar.result()), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param setHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultSetDataObject = function(setHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultSetDataObject(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        setHandler(utils.convReturnListSetDataObject(ar.result()), null);
-      } else {
-        setHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param setHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultSetNullDataObject = function(setHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultSetNullDataObject(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        setHandler(utils.convReturnListSetDataObject(ar.result()), null);
-      } else {
-        setHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param listHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultListEnum = function(listHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultListEnum(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        listHandler(utils.convReturnListSetEnum(ar.result()), null);
-      } else {
-        listHandler(null, ar.cause());
-      }
-    });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param setHandler {function} 
-   */
-  this.methodWithHandlerAsyncResultSetEnum = function(setHandler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      j_testInterface["methodWithHandlerAsyncResultSetEnum(io.vertx.core.Handler)"](function(ar) {
-      if (ar.succeeded()) {
-        setHandler(utils.convReturnListSetEnum(ar.result()), null);
-      } else {
-        setHandler(null, ar.cause());
-      }
-    });
     } else throw new TypeError('function invoked with invalid arguments');
   };
 
@@ -2102,489 +1218,6 @@ var TestInterface = function(j_val) {
         handler(null, ar.cause());
       }
     });
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   @return {Array.<string>}
-   */
-  this.methodWithMapReturn = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      return utils.convReturnMap(j_testInterface["methodWithMapReturn(io.vertx.core.Handler)"](function(jVal) {
-      handler(jVal);
-    }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   @return {Array.<string>}
-   */
-  this.methodWithMapStringReturn = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      return utils.convReturnMap(j_testInterface["methodWithMapStringReturn(io.vertx.core.Handler)"](function(jVal) {
-      handler(jVal);
-    }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   @return {Array.<string>}
-   */
-  this.methodWithMapLongReturn = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      return utils.convReturnMap(j_testInterface["methodWithMapLongReturn(io.vertx.core.Handler)"](function(jVal) {
-      handler(jVal);
-    }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   @return {Array.<string>}
-   */
-  this.methodWithMapIntegerReturn = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      return utils.convReturnMap(j_testInterface["methodWithMapIntegerReturn(io.vertx.core.Handler)"](function(jVal) {
-      handler(jVal);
-    }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   @return {Array.<string>}
-   */
-  this.methodWithMapShortReturn = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      return utils.convReturnMap(j_testInterface["methodWithMapShortReturn(io.vertx.core.Handler)"](function(jVal) {
-      handler(jVal);
-    }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   @return {Array.<string>}
-   */
-  this.methodWithMapByteReturn = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      return utils.convReturnMap(j_testInterface["methodWithMapByteReturn(io.vertx.core.Handler)"](function(jVal) {
-      handler(jVal);
-    }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   @return {Array.<string>}
-   */
-  this.methodWithMapCharacterReturn = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      return utils.convReturnMap(j_testInterface["methodWithMapCharacterReturn(io.vertx.core.Handler)"](function(jVal) {
-      handler(jVal);
-    }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   @return {Array.<string>}
-   */
-  this.methodWithMapBooleanReturn = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      return utils.convReturnMap(j_testInterface["methodWithMapBooleanReturn(io.vertx.core.Handler)"](function(jVal) {
-      handler(jVal);
-    }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   @return {Array.<string>}
-   */
-  this.methodWithMapFloatReturn = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      return utils.convReturnMap(j_testInterface["methodWithMapFloatReturn(io.vertx.core.Handler)"](function(jVal) {
-      handler(jVal);
-    }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   @return {Array.<string>}
-   */
-  this.methodWithMapDoubleReturn = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      return utils.convReturnMap(j_testInterface["methodWithMapDoubleReturn(io.vertx.core.Handler)"](function(jVal) {
-      handler(jVal);
-    }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   @return {Array.<string>}
-   */
-  this.methodWithMapJsonObjectReturn = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      return utils.convReturnMap(j_testInterface["methodWithMapJsonObjectReturn(io.vertx.core.Handler)"](function(jVal) {
-      handler(jVal);
-    }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   @return {Array.<string>}
-   */
-  this.methodWithMapComplexJsonObjectReturn = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      return utils.convReturnMap(j_testInterface["methodWithMapComplexJsonObjectReturn(io.vertx.core.Handler)"](function(jVal) {
-      handler(jVal);
-    }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   @return {Array.<string>}
-   */
-  this.methodWithMapJsonArrayReturn = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      return utils.convReturnMap(j_testInterface["methodWithMapJsonArrayReturn(io.vertx.core.Handler)"](function(jVal) {
-      handler(jVal);
-    }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-   @param handler {function} 
-   @return {Array.<string>}
-   */
-  this.methodWithMapComplexJsonArrayReturn = function(handler) {
-    var __args = arguments;
-    if (__args.length === 1 && typeof __args[0] === 'function') {
-      return utils.convReturnMap(j_testInterface["methodWithMapComplexJsonArrayReturn(io.vertx.core.Handler)"](function(jVal) {
-      handler(jVal);
-    }));
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<string>}
-   */
-  this.methodWithNullMapReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnMap(j_testInterface["methodWithNullMapReturn()"]());
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<string>}
-   */
-  this.methodWithListStringReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return j_testInterface["methodWithListStringReturn()"]();
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<number>}
-   */
-  this.methodWithListLongReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnListSetLong(j_testInterface["methodWithListLongReturn()"]());
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<RefedInterface1>}
-   */
-  this.methodWithListVertxGenReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnListSetVertxGen(j_testInterface["methodWithListVertxGenReturn()"](), RefedInterface1);
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<Object>}
-   */
-  this.methodWithListJsonObjectReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnListSetJson(j_testInterface["methodWithListJsonObjectReturn()"]());
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<Object>}
-   */
-  this.methodWithListComplexJsonObjectReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnListSetJson(j_testInterface["methodWithListComplexJsonObjectReturn()"]());
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<todo>}
-   */
-  this.methodWithListJsonArrayReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnListSetJson(j_testInterface["methodWithListJsonArrayReturn()"]());
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<todo>}
-   */
-  this.methodWithListComplexJsonArrayReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnListSetJson(j_testInterface["methodWithListComplexJsonArrayReturn()"]());
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<Object>}
-   */
-  this.methodWithListDataObjectReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnListSetDataObject(j_testInterface["methodWithListDataObjectReturn()"]());
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<Object>}
-   */
-  this.methodWithListEnumReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnListSetEnum(j_testInterface["methodWithListEnumReturn()"]());
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<string>}
-   */
-  this.methodWithNullListReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return j_testInterface["methodWithNullListReturn()"]();
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<string>}
-   */
-  this.methodWithSetStringReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnSet(j_testInterface["methodWithSetStringReturn()"]());
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<number>}
-   */
-  this.methodWithSetLongReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnListSetLong(j_testInterface["methodWithSetLongReturn()"]());
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<RefedInterface1>}
-   */
-  this.methodWithSetVertxGenReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnListSetVertxGen(j_testInterface["methodWithSetVertxGenReturn()"](), RefedInterface1);
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<Object>}
-   */
-  this.methodWithSetJsonObjectReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnListSetJson(j_testInterface["methodWithSetJsonObjectReturn()"]());
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<Object>}
-   */
-  this.methodWithSetComplexJsonObjectReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnListSetJson(j_testInterface["methodWithSetComplexJsonObjectReturn()"]());
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<todo>}
-   */
-  this.methodWithSetJsonArrayReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnListSetJson(j_testInterface["methodWithSetJsonArrayReturn()"]());
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<todo>}
-   */
-  this.methodWithSetComplexJsonArrayReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnListSetJson(j_testInterface["methodWithSetComplexJsonArrayReturn()"]());
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<Object>}
-   */
-  this.methodWithSetDataObjectReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnListSetDataObject(j_testInterface["methodWithSetDataObjectReturn()"]());
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<Object>}
-   */
-  this.methodWithSetEnumReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnListSetEnum(j_testInterface["methodWithSetEnumReturn()"]());
-    } else throw new TypeError('function invoked with invalid arguments');
-  };
-
-  /**
-
-   @public
-
-   @return {Array.<string>}
-   */
-  this.methodWithNullSetReturn = function() {
-    var __args = arguments;
-    if (__args.length === 0) {
-      return utils.convReturnSet(j_testInterface["methodWithNullSetReturn()"]());
     } else throw new TypeError('function invoked with invalid arguments');
   };
